% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_versions.R
\name{checkAllPackageRequirements}
\alias{checkAllPackageRequirements}
\title{Check that all custom package installation and version requirements are met}
\usage{
checkAllPackageRequirements(pkg_reqs)
}
\arguments{
\item{pkg_reqs}{(character length >= 1) package requirements of same format
as used in the DESCRIPTION DCF files (e.g. "dplyr (>=1.0.1)")}
}
\description{
Check that all custom package installation and version requirements are met
}
