% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_versions.R
\name{checkPackageInstalled}
\alias{checkPackageInstalled}
\title{Check whether a package has been installed}
\usage{
checkPackageInstalled(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Check whether a package has been installed
}
