% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{createReport}
\alias{createReport}
\title{Create IOP Qualification report (PDF)}
\usage{
createReport(
  input_directory,
  settings_directory,
  test_outputs_directory,
  settings_file,
  html_report = FALSE
)
}
\arguments{
\item{input_directory}{Full path to the directory where Rmd file is
stored. [type:character]}

\item{settings_directory}{Full path to the directory where the settings file
is stored. [type:character]}

\item{settings_file}{Name of the settings file (*.yaml) used to generate the
report. [type:character]}

\item{html_report}{Logical flag to enable html report generation. [type:boolean]}

\item{tests_output_directory}{Full path to the directory where test outputs
should be stored (WARNING: any existing files will be deleted). [type:character]}
}
\description{
Create IOP Qualification report (PDF)
}
