% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{customSanitize}
\alias{customSanitize}
\title{Sanitize function which does not sanitize custom coloring of logical values}
\usage{
customSanitize(char, type = "latex")
}
\arguments{
\item{char}{(character length 1) character string to "sanitize"}
}
\description{
Sanitize function which does not sanitize custom coloring of logical values
}
