% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_test_results.R
\name{generateAllTestResults}
\alias{generateAllTestResults}
\title{Generate all test results and return in named list}
\usage{
generateAllTestResults(settings = IOPQUALR_SETTINGS)
}
\arguments{
\item{settings}{(list) IOP qual settings}
}
\description{
Generate all test results and return in named list
}
