% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_help_descriptions.R
\name{getHelpDescription}
\alias{getHelpDescription}
\title{Get help descriptions}
\usage{
getHelpDescription(object = "print")
}
\arguments{
\item{object}{(character length 1) name of object for which to get help description}
}
\description{
Get help descriptions
}
