% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_versions.R
\name{getPkgComparisonOperator}
\alias{getPkgComparisonOperator}
\title{Get the comparison operator from package requirement string (e.g. "==")}
\usage{
getPkgComparisonOperator(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Get the comparison operator from package requirement string (e.g. "==")
}
