% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_versions.R
\name{getPkgName}
\alias{getPkgName}
\title{Get the package name from package requirement string (e.g. "dplyr")}
\usage{
getPkgName(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Get the package name from package requirement string (e.g. "dplyr")
}
