% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_versions.R
\name{getPkgVersion}
\alias{getPkgVersion}
\title{Get the package version from package requirement string (e.g. "1.0-1")}
\usage{
getPkgVersion(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Get the package version from package requirement string (e.g. "1.0-1")
}
