% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{iop_check_all_reqs}
\alias{iop_check_all_reqs}
\title{Check Package Installation and Version Requirements}
\usage{
iop_check_all_reqs(pkg_reqs)
}
\arguments{
\item{pkg_reqs}{A character vector with each element specifying a package
requirement in the format "package (version requirement)". For example,
"dplyr (>=1.0.1)" denotes that the `dplyr` package needs to be installed
and have a version of at least 1.0.1. The vector must have at least one
element.}
}
\value{
A `tibble` with one row per package requirement specified in
`pkg_reqs`. The `tibble` contains the following columns:
  - `requirement`: The package requirement as specified in `pkg_reqs`.
  - `is_installed`: Logical value indicating whether the package is
    installed.
  - `installed_version`: The installed version of the package or `NA` if
    the package is not installed or an error occurred while retrieving
    the version.
  - `all_passed`: Logical value indicating whether the installed version
    meets the specified requirement or `NA` if an error occurred during
    checking.
}
\description{
Ensures that all packages specified in the `pkg_reqs` parameter meet the
custom installation and version requirements. This function is intended to
be used similarly to how package requirements are specified in a
DESCRIPTION file.
}
\examples{
# Example package requirements, replace with actual package names
example_reqs <- c("dplyr (>=1.0.1)", "ggplot2 (>=3.3.3)")
iop_check_all_reqs(example_reqs)
}
