% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{iop_check_installed}
\alias{iop_check_installed}
\title{Check whether a package has been installed}
\usage{
iop_check_installed(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Check whether a package has been installed
}
