% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_validation.R
\name{iop_check_output_dir}
\alias{iop_check_output_dir}
\title{Check if Output Directory Is Valid}
\usage{
iop_check_output_dir(
  input_directory,
  settings_directory,
  test_outputs_directory
)
}
\arguments{
\item{input_directory}{A string representing the path to the input directory.}

\item{settings_directory}{A string representing the path to the settings
directory.}

\item{test_outputs_directory}{A string representing the path to the directory
where test outputs are to be written.}
}
\value{
A logical value indicating whether the test output directory is
valid (`TRUE`) or not (`FALSE`).
If the test output directory does match any of the other directories, or if
it is a subdirectory of them,
then the function will return `FALSE`, otherwise it will return `TRUE`.
}
\description{
This function checks if the test output directory specified does not match
or is not a subdirectory of the input and settings directories.
}
\examples{
\dontrun{
iop_check_output_dir(
  "/my/project/input", "/my/project/settings",
  "/my/project/test-outputs"
)
# Expected output: TRUE or FALSE depending on the context.
}
}
