% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{iop_check_version}
\alias{iop_check_version}
\title{Check that the package version meets the requirement}
\usage{
iop_check_version(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Check that the package version meets the requirement
}
