% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generation.R
\name{iop_color_logical}
\alias{iop_color_logical}
\title{Color false logical values as 'red'}
\usage{
iop_color_logical(d)
}
\arguments{
\item{d}{(data frame)}
}
\description{
Color false logical values as 'red'
}
