% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generation.R
\name{iop_create_report}
\alias{iop_create_report}
\title{Generate an IOP Qualification Report}
\usage{
iop_create_report(
  input_directory,
  settings_directory,
  test_outputs_directory,
  settings_file,
  html_report = FALSE
)
}
\arguments{
\item{input_directory}{The full path to the directory where the RMarkdown
file for the report is stored. This directory should exist and be readable.}

\item{settings_directory}{The full path to the directory where the settings
file is stored. This directory should exist and be readable.}

\item{test_outputs_directory}{The full path to the directory where the test
outputs should be stored. This directory either should already exist or will
be created by the function.}

\item{settings_file}{The name of the settings file (in YAML format) used to
generate the report. It should have a "*.yaml" extension.}

\item{html_report}{A logical flag indicating whether an HTML report should
be generated instead of a PDF. Defaults to FALSE, meaning a PDF report is
 generated by default.}
}
\value{
The function does not return a value but generates a report file in
 the specified output format and location.
}
\description{
Creates a report in either PDF or HTML format based on specified settings.
 The function relies on `pdfcrop` if a PDF report is required (unless
 `html_report` is TRUE), and the settings for the report are read from a
 YAML file. The function performs checks to ensure all directories and files
 are in place before proceeding with the report generation. If required
 checks fail, it will stop and provide an informative error message.
}
\examples{
\dontrun{
iop_create_report(
  input_directory = "/path/to/input",
  settings_directory = "/path/to/settings",
  test_outputs_directory = "/path/to/test/output",
  settings_file = "settings.yaml",
  html_report = FALSE
)
}
}
