% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generation.R
\name{iop_custom_add_row}
\alias{iop_custom_add_row}
\title{Add statement to final row of multi-page table}
\usage{
iop_custom_add_row()
}
\description{
Add statement to final row of multi-page table
}
