% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generation.R
\name{iop_custom_sanitize}
\alias{iop_custom_sanitize}
\title{Sanitize function which does not sanitize custom coloring of logical values}
\usage{
iop_custom_sanitize(char, type = "latex")
}
\arguments{
\item{char}{(character length 1) character string to "sanitize"}

\item{type}{(character length 1) type of sanitization to apply,
defaults to "latex"}
}
\description{
Sanitize function which does not sanitize custom coloring of logical values
}
