% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results_management.R
\name{iop_gen_test_results}
\alias{iop_gen_test_results}
\title{Generate all test results and return in named list}
\usage{
iop_gen_test_results(settings)
}
\arguments{
\item{settings}{(list) IOP qual settings}
}
\description{
Generate all test results and return in named list
}
\examples{
\dontrun{
settings_directory <- "path/to/settings"
settings_file <- "settings_file_name"
settings <- iop_load_input_file(file.path(settings_directory, settings_file))
results <- iop_gen_test_results(settings)
}
}
