% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{iop_get_comp_operator}
\alias{iop_get_comp_operator}
\title{Get the comparison operator from package requirement string (e.g. "==")}
\usage{
iop_get_comp_operator(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Get the comparison operator from package requirement string (e.g. "==")
}
