% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{iop_get_help_desc}
\alias{iop_get_help_desc}
\title{Get help descriptions}
\usage{
iop_get_help_desc(object = "print")
}
\arguments{
\item{object}{(character length 1) name of object for which to get help
description}
}
\description{
Get help descriptions
}
