% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{iop_get_pkg_name}
\alias{iop_get_pkg_name}
\title{Get the package name from package requirement string (e.g. "dplyr")}
\usage{
iop_get_pkg_name(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Get the package name from package requirement string (e.g. "dplyr")
}
