% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{iop_get_pkg_version}
\alias{iop_get_pkg_version}
\title{Get the package version from package requirement string (e.g. "1.0-1")}
\usage{
iop_get_pkg_version(pkg_req)
}
\arguments{
\item{pkg_req}{(character length >= 1) package requirement strings}
}
\description{
Get the package version from package requirement string (e.g. "1.0-1")
}
