% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_management.R
\name{iop_input_helper}
\alias{iop_input_helper}
\title{Interactive Helper to Write and Save an Yaml Input File}
\usage{
iop_input_helper()
}
\value{
Invisibly returns `NULL`. A YAML file with the collected settings is
 created and saved in the specified location.
}
\description{
This function provides an interactive session to the user for setting up
  a YAML input file required by the IOPQUALR package. It guides the user
  through a series of questions to capture necessary information, such as
  description, author, R version, targeted packages, and test directories.
  The collected data is then saved into a YAML file at a chosen location.
}
\details{
The function is interactive and requires the user to input answers to
 prompted questions in the console.
 The user is guided to provide the following information:

- A descriptive text for the report.
- The name of the report's author or the responsible company.
- The targeted version of R for the tests.
- A list of packages to be tested, along with their specific versions.
- The path to a custom test directory.
- Lists of packages for performance and operational tests.
- A selection of test types to perform (examples, vignettes, tests).

After gathering all the inputs, the function will ask the user for a
directory and file name to save the generated YAML file, which contains
all the configured settings.
}
\examples{
\dontrun{
# Run the function to start the interactive input helper
iop_input_helper()
}
}
