% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_management.R
\name{iop_load_input_file}
\alias{iop_load_input_file}
\title{Load a YAML Test Input File}
\usage{
iop_load_input_file(
  input_file_path = system.file("extdata", "sample_input.yaml", package = "iopqualr"),
  tests_output_dir = "/home/iopqualr/"
)
}
\arguments{
\item{input_file_path}{A string representing the full file path to the YAML
input file. If not provided, the default sample input file from the
package's `extdata` directory is used.}

\item{tests_output_dir}{A string representing the full path to the Youtput
dir for tests execution}
}
\value{
Returns a list structure containing the contents of the YAML file.
}
\description{
This function reads a YAML file from the specified path and loads it into R.
The path provided should point to a YAML file. By default, the function looks
for a file named `sample_input.yaml` within the `extdata` directory of the
`iopqualr` package.
}
\examples{
\dontrun{
# Load the default sample input YAML file
data <- iop_load_input_file()

# Specify a path to a custom YAML file
custom_path <- "/path/to/your/input.yaml"
data <- iop_load_input_file(custom_path, tests_output_path)
}
}
