% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results_management.R
\name{iop_run_basic_tests}
\alias{iop_run_basic_tests}
\title{Run basic installed test}
\usage{
iop_run_basic_tests(scope = "both")
}
\arguments{
\item{scope}{(character length 1) which tests to run, should be one of
"basic"" "devel", "both", "internet"}
}
\description{
Run basic installed test
}
