% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results_management.R
\name{iop_summarise_results}
\alias{iop_summarise_results}
\title{Summarize all the test results into one table}
\usage{
iop_summarise_results(results = iop_gen_test_results())
}
\arguments{
\item{results}{(list of data frames) as created by `iop_gen_test_results()`}
}
\description{
Summarize all the test results into one table
}
