% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results_management.R
\name{iop_test_base_recomm}
\alias{iop_test_base_recomm}
\title{Run all tests for base and recommended packages}
\usage{
iop_test_base_recomm(
  tests_to_run = list(list(scope = "base", types = "examples"), list(scope = "base",
    types = "vignettes"), list(scope = "base", types = "tests"), list(scope =
    "recommended", types = "examples"), list(scope = "recommended", types = "vignettes"),
    list(scope = "recommended", types = "tests")),
  output_directory = settings[["tests_output_dir"]],
  ...
)
}
\arguments{
\item{tests_to_run}{(list of lists with "scope" and "type" elements)}

\item{output_directory}{The file path where test results should be saved.
By default the current working directory is used}

\item{...}{The additional arguments that are passed on to the
`tools::testInstalledPackages()` function. This allows for
customizing the behavior of the test execution with extra
arguments such as `filter`, `stopOnError`, and others that are}
}
\description{
Run all tests for base and recommended packages
}
