% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results_management.R
\name{iop_test_custom_pkgs}
\alias{iop_test_custom_pkgs}
\title{Test non-base, non-recommended packages using their included tests}
\usage{
iop_test_custom_pkgs(
  tests_to_run,
  output_directory = settings[["tests_output_dir"]],
  ...
)
}
\arguments{
\item{tests_to_run}{A list of lists containing the packages to be tested
and the types of tests to be run. Each sublist should have elements "pkg"
(a character string specifying the name of the package to be tested)
and "types" (a character vector specifying which types of tests to run).}

\item{output_directory}{A character string specifying the directory where
test results will be saved. If not provided, defaults to the current
working directory}

\item{...}{Additional arguments passed to `tools::testInstalledPackage`.}
}
\value{
A `data.frame` that summarizes the test results for each package.
Each row in the data frame will have the package name, the types of tests
that were run, and a logical indicator of whether the tests passed.
Returns `NULL` if no tests are run.
}
\description{
This function takes a list of non-base, non-recommended R packages and
runs the tests included with each package. It is used to ensure that
custom-installed packages are working as expected. The results of the tests
are output to a specified directory.
}
\examples{
\dontrun{
# Assuming you have a customized list of tests in the format required:
tests_list <- list(
  list(pkg = "ggplot2", types = c("tests")),
  list(pkg = "dplyr", types = c("tests", "vignettes"))
)
result <- iop_test_custom_pkgs(tests_to_run = tests_list, out_dir)

# To use the default settings for tests to run and output directory:
settings_directory <- "path/to/settings"
settings_file <- "settings_file_name"
output_directory <- "path/to/out/folder"
settings <- iop_load_input_file(
  file.path(settings_directory, settings_file),
  out_path
)
tests_list <- iop_get_custom_tests(settings)
result <- iop_test_custom_pkgs(
  tests_to_run = tests_list,
  output_directory = output_dir
)
}
}
