% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_inputs.R
\name{loadInputFile}
\alias{loadInputFile}
\title{Load a YAML test input file}
\usage{
loadInputFile(
  input_file_path = system.file("extdata", "sample_input.yaml", package = "iopqualr")
)
}
\arguments{
\item{input_file_path}{(character length 1) full file path to input file}
}
\description{
Load a YAML test input file
}
