% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_installed_basic.R
\name{runBasicInstalledTest}
\alias{runBasicInstalledTest}
\title{Run basic installed test}
\usage{
runBasicInstalledTest(scope = "both")
}
\arguments{
\item{scope}{(character length 1) which tests to run}
}
\description{
Run basic installed test
}
