% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_custom.R
\name{runCustomTests}
\alias{runCustomTests}
\title{Custom operation and performance tests}
\usage{
runCustomTests(packages = NULL, type = NULL, tests_path = NULL)
}
\arguments{
\item{packages}{(character length >= 1)}

\item{type}{(character length 1) either "operational" or "performance", the
test "type"}

\item{tests_path}{(character length 1) the location of the tests}
}
\description{
Custom operation and performance tests
}
