% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tests.R
\name{summariseTestResults}
\alias{summariseTestResults}
\title{Summarise all the test results into one table}
\usage{
summariseTestResults(test_results = generateAllTestResults())
}
\arguments{
\item{test_results}{(list of data frames) as created by `generateAllTestResutls`}
}
\description{
Summarise all the test results into one table
}
