% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_base_recommended_packages.R
\name{testBaseAndRecommendedPackages}
\alias{testBaseAndRecommendedPackages}
\title{Run all tests for base and recommended packages}
\usage{
testBaseAndRecommendedPackages(
  tests_to_run = list(list(scope = "base", types = "examples"), list(scope = "base",
    types = "vignettes"), list(scope = "base", types = "tests"), list(scope =
    "recommended", types = "examples"), list(scope = "recommended", types = "vignettes"),
    list(scope = "recommended", types = "tests")),
  output_directory = TESTS_OUTPUT_DIRECTORY,
  ...
)
}
\arguments{
\item{tests_to_run}{(list of lists with "scope" and "type" elements)}
}
\description{
Run all tests for base and recommended packages
}
