% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_installed_packages.R
\name{testCustomInstalledPackages}
\alias{testCustomInstalledPackages}
\title{Test on non-base, non-recommended packages by their included tests}
\usage{
testCustomInstalledPackages(
  tests_to_run = CUSTOM_TESTS_TO_RUN(IOPQUALR_SETTINGS),
  output_directory = TESTS_OUTPUT_DIRECTORY,
  ...
)
}
\arguments{
\item{tests_to_run}{(list of lists with elements "pkg" and "types")}
}
\description{
Test on non-base, non-recommended packages by their included tests
}
